Externalize number of processed cases counter
Move junit test to integration test to cover the new property
Add logging to know when tokes are renewed
Fix health endpoint configuration
