+++ src/test/java/uk/gov/hmcts/reform/sscs/services/CaseLoaderServiceTest.java
@@ -1,11 +1,6 @@
 package uk.gov.hmcts.reform.sscs.services;
 
 import static com.google.common.collect.Lists.newArrayList;
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.eq;
@@ -15,13 +10,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
 import javax.xml.stream.XMLStreamException;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
@@ -104,76 +96,6 @@ public class CaseLoaderServiceTest {
 
         verify(xmlValidator, times(2)).validateXml(file);
         verify(sftpSshService, times(2)).move(file, true);
-    }
-
-
-    @Test
-    public void givenItHasProcessed100Cases_shouldRenewAuthenticationTokens() {
-        IdamTokens idamTokens = IdamTokens.builder()
-            .idamOauth2Token("oAuth2Token")
-            .serviceAuthorization("s2sToken")
-            .userId("16")
-            .build();
-
-        IdamTokens idamTokens2 = IdamTokens.builder()
-            .idamOauth2Token("oAuth2Token2")
-            .serviceAuthorization("s2sToken2")
-            .userId("16")
-            .build();
-
-        IdamTokens idamTokens3 = IdamTokens.builder()
-            .idamOauth2Token("oAuth2Token3")
-            .serviceAuthorization("s2sToken3")
-            .userId("16")
-            .build();
-
-        given(idamService.getIdamTokens())
-            .willReturn(idamTokens)
-            .willReturn(idamTokens2)
-            .willReturn(idamTokens3);
-
-        given(transformService.transform(inputStream)).willReturn(buildCaseListWithGivenNumberOfElements(201));
-
-        given(searchCcdCaseService.findCaseByCaseRefOrCaseId(any(SscsCaseData.class), any(IdamTokens.class)))
-            .willReturn(null)
-            .willReturn(SscsCaseDetails.builder().build());
-
-        caseLoaderService.process();
-
-        then(idamService).should(times(3)).getIdamTokens();
-
-        ArgumentCaptor<IdamTokens> idamTokensArgumentCaptor = ArgumentCaptor.forClass(IdamTokens.class);
-        then(searchCcdCaseService).should(times(201))
-            .findCaseByCaseRefOrCaseId(eq(caseData), idamTokensArgumentCaptor.capture());
-        List<IdamTokens> idamTokensValues = idamTokensArgumentCaptor.getAllValues();
-
-        long actualNumberOfIdamTokens = idamTokensValues.stream()
-            .filter(token -> token == idamTokens)
-            .count();
-        assertThat(actualNumberOfIdamTokens, is(equalTo(100L)));
-
-        long actualNumberOfIdamTokens2 = idamTokensValues.stream()
-            .filter(token -> token == idamTokens2)
-            .count();
-        assertThat(actualNumberOfIdamTokens2, is(equalTo(100L)));
-
-        long actualNumberOfIdamTokens3 = idamTokensValues.stream()
-            .filter(token -> token == idamTokens3)
-            .count();
-        assertThat(actualNumberOfIdamTokens3, is(equalTo(1L)));
-
-        then(xmlValidator).should(times(2)).validateXml(file);
-        then(sftpSshService).should(times(2)).move(file, true);
-    }
-
-    private List<SscsCaseData> buildCaseListWithGivenNumberOfElements(int elements) {
-        List<SscsCaseData> cases = new ArrayList<>(elements);
-        int counter = 0;
-        while (counter < elements) {
-            cases.add(caseData);
-            counter++;
-        }
-        return cases;
     }
 
     @Test
